change 'import' to 'java_import' to avoid jruby/rake conflict
both jruby and rake monkey-patch Object.import (see
http://jira.codehaus.org/browse/JRUBY-3171 and
http://www.mail-archive.com/rake-devel@rubyforge.org/msg00525.html for
more info).  this leads to problems when a ruby class tries to import
a java class.  i'm not sure why this didn't fail in rails 2.x, but it
definitely fails in rails 3.

the short-term solution is to use java_import to import java code.
longer-term, rake will put 'import' into a namespace.